[B] feat : Notice : 공지사항 목록 조회 API에 read, readCount 필드 추가 #26

+++ backend/src/main/java/ssu/groupstudy/domain/comment/repository/CommentRepository.java
@@ -18,5 +18,5 @@ public interface CommentRepository extends JpaRepository<Comment, Long> {
 
 
     @Query("SELECT COUNT(c) FROM Comment c WHERE c.notice = :notice AND c.deleteYn = 'N'")
-    Long countCommentByNotice(Notice notice);
+    int countCommentByNotice(Notice notice);
 }
